pyproject.toml
[project]
name = "identificacion-de-compromisos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain-community>=0.3.18",
    "langchain>=0.3.19",
    "pillow>=11.1.0",
    "pydantic>=2.10.6",
    "pymupdf>=1.25.3",
    "python-dotenv>=1.0.1",
    "ruff>=0.9.8",
    "tqdm>=4.67.1",
    "pandas>=2.2.3",
    "openpyxl>=3.1.5",
    "langchain-openai>=0.3.7",
    "langchain-google-genai>=2.0.11",
    "pymupdf4llm>=0.0.17",
    "tabulate>=0.9.0",
    "mistralai>=1.5.1",
    "fastparquet>=2024.11.0",
    "pyarrow>=19.0.1",
    "gradio>=5.20.1",
    "matplotlib>=3.10.1",
    "seaborn>=0.13.2",
    "vizro>=0.1.34",
]
